+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/features/code/cat/ClassModifierFeature.java
@@ -1,50 +1,51 @@
-package de.fraunhofer.iem.swan.features.code;
+package de.fraunhofer.iem.swan.features.code.cat;
 
-import de.fraunhofer.iem.swan.data.Category;
 import de.fraunhofer.iem.swan.data.Method;
-import de.fraunhofer.iem.swan.features.code.type.WeightedFeature;
+import de.fraunhofer.iem.swan.features.code.FeatureResult;
+import de.fraunhofer.iem.swan.features.code.ICodeFeature;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class ClassAccessModifierFeature extends WeightedFeature implements IFeatureNew{
+public class ClassModifierFeature implements ICodeFeature {
 
     private Modifier modifier;
     private FeatureResult featureResult;
     private ArrayList<String> featureValues;
 
-    public ClassAccessModifierFeature() {
+    public enum Modifier{
+        STATIC, ABSTRACT, FINAL, DEFAULT
+    }
+
+    public ClassModifierFeature() {
         this.featureResult = new FeatureResult();
     }
 
-    public enum Modifier{
-        PRIVATE, PUBLIC, PROTECTED, DEFAULT,
-    }
-
     @Override
-    public FeatureResult applies(Method method, Category category) {
-        if(method.getSootMethod().getDeclaringClass().isPublic()){
-            this.modifier = Modifier.PUBLIC;
-        } else if (method.getSootMethod().getDeclaringClass().isPrivate()) {
-            this.modifier = Modifier.PRIVATE;
-        } else if (method.getSootMethod().getDeclaringClass().isProtected()) {
-            this.modifier = Modifier.PROTECTED;
+    public FeatureResult applies(Method method) {
+        if(method.getSootMethod().getDeclaringClass().isFinal()){
+            this.modifier = Modifier.FINAL;
+        } else if (method.getSootMethod().getDeclaringClass().isStatic()) {
+            this.modifier = Modifier.STATIC;
+        } else if (method.getSootMethod().getDeclaringClass().isAbstract()) {
+            this.modifier = Modifier.ABSTRACT;
         } else {
             this.modifier = Modifier.DEFAULT;
         }
         this.featureResult.setStringValue(String.valueOf(this.modifier));
         return this.featureResult;
     }
-
     @Override
     public FeatureType getFeatureType() {
         return FeatureType.CATEGORICAL;
     }
 
+
+
     @Override
     public String toString(){
-        return "ClassAccessModifier";
+        return "ClassModifier";
     }
 
     @Override
