+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/features/code/cat/ReturnTypeFeature.java
@@ -1,15 +1,13 @@
-package de.fraunhofer.iem.swan.features.code;
+package de.fraunhofer.iem.swan.features.code.cat;
 
-import de.fraunhofer.iem.swan.data.Category;
 import de.fraunhofer.iem.swan.data.Method;
-import de.fraunhofer.iem.swan.features.code.type.WeightedFeature;
-
+import de.fraunhofer.iem.swan.features.code.FeatureResult;
+import de.fraunhofer.iem.swan.features.code.ICodeFeature;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class ReturnTypeFeature extends WeightedFeature implements IFeatureNew {
+public class ReturnTypeFeature implements ICodeFeature {
 
     private Values category;
 
@@ -25,19 +23,19 @@ public class ReturnTypeFeature extends WeightedFeature implements IFeatureNew {
     }
 
     @Override
-    public FeatureResult applies(Method method, Category category){
+    public FeatureResult applies(Method method){
         switch (method.getReturnType()){
             case "java.lang.CharSequence":
             case "java.lang.String":
                 this.category = Values.String;
                 break;
-
+            //TODO maybe split the categories for numerical, bytes and boolean
             case "byte[]":
             case "int":
             case "boolean":
                 this.category = Values.Native;
                 break;
-
+            //TODO we could find a way to split these into more categories
             case "User":
             case "Credential":
             case "java.sql.ResultSet":
